Edit client added

+++ ClientManager.DAL/Interfaces/IRepository.cs
@@ -6,10 +6,10 @@ namespace LearningPortal.DAL.Interfaces
     public interface IRepository<TEntity> where TEntity : class
     {
         IQueryable<TEntity> Get();
-        TEntity GetById(int id);
+        TEntity GetById(long id);
         void Add(TEntity entity);
         void AddRange(IList<TEntity> values);
-        void Delete(int id);
+        void Delete(long id);
         void Delete(TEntity entity);
         void Update(TEntity entity);
         void SaveChanges();
